Move functions to API
