+++ seed_data/make_requests.py
@@ -1,7 +1,4 @@
-import requests
 from requests import request
-
-from conf.settings import env
 
 
 def make_request(method, base_url, url, headers, body=None, files=None):
@@ -14,33 +11,3 @@ def make_request(method, base_url, url, headers, body=None, files=None):
         raise Exception("bad response: " + response.text)
 
     return response
-
-
-def get(appended_address, headers):
-    return requests.get(env("LITE_API_URL") + appended_address, headers=headers)
-
-
-def post(appended_address, json, headers):
-    return requests.post(
-        env("LITE_API_URL") + appended_address,
-        json=json,
-        headers=headers
-    )
-
-
-def put(appended_address, json, headers):
-    if not appended_address.endswith("/"):
-        appended_address = appended_address + "/"
-
-    return requests.put(
-        env("LITE_API_URL") + appended_address,
-        json=json,
-        headers=headers
-    )
-
-
-def delete(appended_address, headers):
-    return requests.delete(
-        env("LITE_API_URL") + appended_address,
-        headers=headers,
-    )
